[Sparrow mahjong] Add baseline (#924)

+++ workspace/ppo_training_haiku/vis.py
@@ -4,7 +4,7 @@ import haiku as hk
 import pgx
 import time
 import pickle
-from ppo import ActorCritic
+from ppo_multi import ActorCritic
 import distrax
 import argparse
 TRUE = jnp.bool_(True)
@@ -19,7 +19,7 @@ def visualize(forward_pass, model, env,  rng_key):
     states.append(state)
     step_fn = jax.jit(jax.vmap(env.step))
     while not state.terminated.all():
-        (logits, value), _ = forward_pass.apply(model_params,model_state, state.observation, is_eval=True)
+        (logits, value), _ = forward_pass.apply(model_params,model_state, state.observation.astype(jnp.float32), is_eval=True)
         logits = logits +  jnp.finfo(jnp.float64).min * (~state.legal_action_mask)
         pi = distrax.Categorical(logits=logits)
         rng_key, _rng = jax.random.split(rng_key)
